addressed feedback (refernced the right class types, filled in info gaps in the method description, removed BranchLastAttributedTouchData object because it doesn't do anything and replaced it with a standalone ServerResuestGetLATD class, this way we we also don't end up making a request from a constructor class
