+++ Branch-SDK/src/main/java/io/branch/referral/Branch.java
@@ -24,8 +24,7 @@ import android.view.View;
 import io.branch.referral.Defines.PreinstallKey;
 import io.branch.referral.util.BranchCrossPlatformId;
 import io.branch.referral.util.BranchCrossPlatformId.BranchCrossPlatformIdListener;
-import io.branch.referral.util.BranchLastAttributedTouchData;
-import io.branch.referral.util.BranchLastAttributedTouchData.BranchLastAttributedTouchDataListener;
+import io.branch.referral.ServerRequestGetLATD.BranchLastAttributedTouchDataListener;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -1829,7 +1828,8 @@ public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserv
     /**
      * Gets all available cross platform ids.
      *
-     * @param callback An instance of {@link io.branch.referral.util.BranchLastAttributedTouchData} to callback with cross platform ids
+     * @param callback An instance of {@link io.branch.referral.util.BranchCrossPlatformId.BranchCrossPlatformIdListener}
+     *                to callback with cross platform ids
      *
      */
     public void getCrossPlatformIds(BranchCrossPlatformIdListener callback) {
@@ -1839,24 +1839,32 @@ public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserv
     /**
      * Gets the available last attributed touch data.
      *
-     * @param callback An instance of {@link io.branch.referral.util.BranchLastAttributedTouchData} to callback with last attributed touch data
+     * @param callback An instance of {@link io.branch.referral.ServerRequestGetLATD.BranchLastAttributedTouchDataListener}
+     *                 to callback with last attributed touch data
      *
      */
     public void getLastAttributedTouchData(BranchLastAttributedTouchDataListener callback) {
         prefHelper_.setLATDAttributionWindow(null);
-        new BranchLastAttributedTouchData(callback, context_);
+        if (context_ != null) {
+            handleNewRequest(new ServerRequestGetLATD(context_, Defines.RequestPath.GetLATD.getPath(), callback));
+        }
     }
 
     /**
      * Gets the available last attributed touch data with a custom set attribution window.
      *
-     * @param callback An instance of {@link io.branch.referral.util.BranchLastAttributedTouchData} to callback with last attributed touch data
-     * @param attributionWindow A {@link int}
+     * @param callback An instance of {@link io.branch.referral.ServerRequestGetLATD.BranchLastAttributedTouchDataListener}
+     *                to callback with last attributed touch data
+     * @param attributionWindow An {@link int} to bound the the window of time in days during which
+     *                          the attribution data is considered valid. Note that, server side, the
+     *                          maximum value is 90.
      *
      */
     public void getLastAttributedTouchData(BranchLastAttributedTouchDataListener callback, int attributionWindow) {
         prefHelper_.setLATDAttributionWindow(attributionWindow);
-        new BranchLastAttributedTouchData(callback, context_);
+        if (context_ != null) {
+            handleNewRequest(new ServerRequestGetLATD(context_, Defines.RequestPath.GetLATD.getPath(), callback));
+        }
     }
 
     /**
